% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_data}
\alias{bc_data}
\title{Breast cancer data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 59 rows and 737 columns.
}
\source{
\url{https://portal.gdc.cancer.gov}
}
\usage{
bc_data
}
\description{
Beta-values of 59 informative sites from 737 samples from TCGA.
}
\keyword{datasets}
