% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc_subtypes}
\alias{bc_subtypes}
\title{PAM50 subtypes of breast cancer samples}
\format{
An object of class \code{character} of length 737.
}
\source{
\url{https://bioconductor.org/packages/release/bioc/vignettes/TCGAbiolinks/inst/doc/subtypes.html}
}
\usage{
bc_subtypes
}
\description{
PAM50 subtypes of 737 BC samples from TCGA reported in \code{bc_data}.
}
\keyword{datasets}
