% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmnet_bc_model}
\alias{glmnet_bc_model}
\title{GLMNET breast cancer model}
\format{
An object of class \code{cv.glmnet} of length 12.
}
\usage{
glmnet_bc_model
}
\description{
GLMNET model generated with \code{cv.glmnet} to classify breast cancer samples into three PAM50 samples (Luminal A/B, HER2-e, Basal-like).
}
\keyword{datasets}
