% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmnet_multicancer_model}
\alias{glmnet_multicancer_model}
\title{GLMNET multicancer cancer model}
\format{
An object of class \code{cv.glmnet} of length 12.
}
\usage{
glmnet_multicancer_model
}
\description{
GLMNET model generated with \code{cv.glmnet} to classify cancer samples into 14 possible cancer-types:
"BLCA"      = 'bladder"
"BRCA"      = 'breast"
"COAD"      = 'colon"
"ESCASC"    = 'esophagus"
"HNSC"      = 'oral|neck"
"KIRC|KIRP" = 'kidney"
"LIHC"      = 'liver"
"LUAD|LUSC" = 'lung"
"PAAD"      = 'pancreas"
"PRAD"      = 'prostate"
"THCA"      = 'thyroid"
"UCEC"      = 'uterus"
}
\keyword{datasets}
