% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_glmnet_model.R
\name{make_glmnet_model}
\alias{make_glmnet_model}
\title{Make a classification model using GLMNET}
\usage{
make_glmnet_model(x, y, k = 5, cores = 1, seed = NULL)
}
\arguments{
\item{x}{A matrix of beta-values (sitesXsamples)to train the model.}

\item{y}{A vector reporting the classes (tumor type/subtype) of the samples.}

\item{k}{An integer. Number of folds for cross-validation (default=5).}

\item{cores}{An integer. Number of parallel processes (default=1).}

\item{seed}{A number. Set the seed for the pseudorandom number generator
(default to random) for reproducible analyses.}
}
\description{
Make a classification model using GLMNET
}
\examples{
glmnet_model <- make_glmnet_model(bc_data, bc_subtypes, cores=20)

}
