% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpc_bc_model}
\alias{rpc_bc_model}
\title{RPC breast cancer model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 59 rows and 3 columns.
}
\usage{
rpc_bc_model
}
\description{
RPC model generated taking median beta-values of informative sites for each PAM50 subtype (Luminal A/B, HER2-e, Basal-like).
}
\keyword{datasets}
