% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpc_multicancer_model}
\alias{rpc_multicancer_model}
\title{RPC multicancer cancer model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 278 rows and 15 columns.
}
\usage{
rpc_multicancer_model
}
\description{
RPC model generated taking median beta-values of informative sites for 14 cancer-types:
"BLCA"      = 'bladder"
"BRCA"      = 'breast"
"COAD"      = 'colon"
"ESCASC"    = 'esophagus"
"HNSC"      = 'oral|neck"
"KIRC|KIRP" = 'kidney"
"LIHC"      = 'liver"
"LUAD|LUSC" = 'lung"
"PAAD"      = 'pancreas"
"PRAD"      = 'prostate"
"THCA"      = 'thyroid"
"UCEC"      = 'uterus"
}
\keyword{datasets}
